Medium progress update.
Added custom background image utility class and height and colors to Tailwind.
Refactored the Search component, and pulled out and created the main Search Form and it's
functionality out of Search. Now it is a reusable component in the HeroSection and in the Search component.
Any of it's use now redirects to search component and shows results based on queries.
Input still not functioning as should.
Added concurrently dual start for backend and frontend.
Added hero image bg.