PROGRESS UPDATE:
Git add before pulls produced losses in changes, so they aren't claimed as "changed" or "modified" within this commit even though they occurred.

FRONTEND:

CreatePost redux hook up was created, along with Reducer, actions, and component. Have to add additional logic both on the frontend and backend to handle the links that can't be scraped (and provide a list on the frontend for this.)

BACKEND:

- Added npm package Cheerio for html parsing of user-entered websites (not including from the sourcing of reddit sites. Strictly for on platform post creation.)
- Expanded the Post object class, and refactored the data that would be entered by a user to populate entries for the posts table. Now, the title and image is scraped from a website through (with the exception of sites like: Amazon, Walmart AMD, Bestbuy, and zotacstore (for now)) through Cheerio.
- Certain sites are not able to be scraped (like those stated above), so these will be optionally selected by a user to enter title if the site is one of those listed above.